backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
numpy==1.26.4
openai==1.109.0
packaging==25.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.3.0
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
starlette==0.38.6
tiktoken==0.7.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
